refactor(threads): useThreadActions.helpers 拆四族 + re-export 桶
- 2646→243 行：recovery 1094 / engineSummaries 988 / continuityMerge 260 / rewind 88
- 主文件保留尾部 scan/snapshot 工具 ~230 行 + 四族 export * 桶，helpers.test.ts 零改动
- normalizeGeminiSessionSummary 归 engineSummaries；RewindSupportedEngine 归 rewind
- baseline 重生成 + 新文件登记豁免
验证：tsc 零 error；helpers.test.ts 60 passed；其余 useThreadActions 测试失败与基线一致（tauri mock 既有问题）
